🎨 Move modint-related headers to subdirectory
