+++ template/fps/integral.hpp
@@ -1,6 +1,6 @@
 #pragma once
 #include "base_template.hpp"
-#include "modint_petit_p.hpp"
+#include "modint/modint_petit_p.hpp"
 
 // FPS 積分 O(N) 定数項は 0 とする
 template <ll pdiv, ll prim>
